feat: Add DiffVals func
This allows you to check two values of any type for equality and returns an
error if they are different. It is very like the standard reflect.DeepEqual
func but it returns an error and the error tells you what the difference is
rather than just a bool. Another useful feature is that it allows you to
ignore fields in structs if you expect them to differ.